Coins/requirements.txt
python-dotenv
requests
pymongo
streamlit
pandas


